♻️ Refactor: change route name (#20)
- 기존의 home route명을 login으로 변경
- 기존의 myPage route명을 home으로 변경
