+++ src/components/form/SignUpForm.vue
@@ -24,7 +24,7 @@
           </v-card-text>
           <v-card-actions class="flex-column pa-0">
             <v-btn class="emphasis mb-16" size="x-large" flat @click="signUp">가입하기</v-btn>
-            <router-link :to="{ name: 'home' }">
+            <router-link :to="{ name: 'login' }">
               <v-btn class="ma-0" size="large" prepend-icon="mdi-chevron-left" flat>로그인 화면으로 돌아가기</v-btn>
             </router-link>
           </v-card-actions>
@@ -33,7 +33,7 @@
           <v-card-title class="mb-5">공책에 오신 것을 환영해요!</v-card-title>
           <v-card-text class="mb-10 pa-0">회원 가입 후 자동으로 로그인되었어요.</v-card-text>
           <v-card-actions class="pa-0">
-            <router-link :to="{ name: 'myPage' }">
+            <router-link :to="{ name: 'home' }">
               <v-btn class="emphasis" size="large">나만의 독서 기록 채우러 가기</v-btn>
             </router-link>
           </v-card-actions>
