+++ src/views/Login.vue
@@ -1,0 +1,10 @@
+<template>
+  <LoginForm />
+</template>
+
+<script setup>
+import LoginForm from '@/components/form/LoginForm.vue';
+</script>
+
+<style scoped>
+</style>
